Initial commit
We throw together an initial setup that doesn't do anything yet.

We're going to try throwing a bunch of PureScript tooling into this
repository. Hopefully, we'll have an alright time with it.

This is also our first foray into using the newer Haskell IDE tooling.
It's just as hard to get started, but it seems to work pretty well once
it's up and running. Hopefully, it pays off in the long run.