Extract JavaScript files to their own data type
We can group all of the JavaScript files data together.

We follow the idea of what we did for the externs files for the
JavaScript files. Since we only have one JavaScript file at the moment,
it might seem like overkill. But, we're about to get kind of complex
with the FFI files as well. This change will start to pay off soon.