Start on `rules_purescript`
We want a better way to use `purs-compile-module` from `bazel`.

We're following Jay Conrod's series on writing bazel rules:
https://jayconrod.com/posts/106/writing-bazel-rules--simple-binary-rule.
It's not clear yet if we're going to do everything in the series. But,
this seems like one of the most complete walkthroughs of writing `bazel`
rules from scratch.

We start by throwing together a simple "binary" rule for PureScript. We
end up depending on `rules_nodejs` since we need some way to run the
program.

We're mimicking the layout for a rules repo:
https://docs.bazel.build/versions/master/skylark/deploying.html. It's
not quite the same since we're not doing this at the top-level. But
aside from that, we try to follow the recommendation. If this becomes a
usable thing and we need it to be in a separate repository, we can split
it out with git.

In any case, we have something that works. We'll continue as things
progress.