Add `purescript_library` rule
This rule allows creating dependencies for PureScript.

The first thing to note is that much like the `purescript_binary` rule,
the `purescript_library` rule corresponds to a PureScript module. It
does don't correspond to a PureScript "package." This is largely because
there's no actual thing as a PureScript "package." That's a convention
brought on by the vast majority of people that use PureScript. We can
add a rule (or macro) that helps with working with PureScript
"package"s, but this `purescript_library` rule isn't the place for it.

As far as what this `purescript_library` rule does, it acts like other
bazel `*_library` rules. I.e. it provides a way to decompose a bazel
package into smaller units. We need to do quite a bit of work to make it
all work seemlessly. Since we're still suck with `purs-compile-module`
not knowing how to work with direct dependencies only, we have to
flatten each depset every time we want to use it. This is a huge
performance issue.

The right way to deal with the flattening of depsets is to fix
`purs-compile-module` to be able to work with direct dependencies only.
It's entirely possible that we _could_ fix it on the bazel-side. But,
it's still a naïve model to have a compiler that requires every single
transitive dependency in order to compile anything.

We also create a provider in order to pass along the information. It
seems like we got all of the information we need, but we might need a
bit more.

In any case, this seems to work out okay. And it means we can build any
PureScript project. Actually attempting to do that by hand would be
incredibly tedious, so there's definitely more we can improve as far as
the workflow goes.

We show how we can use the rules in the `//examples/purs-compile-module`
package. It's quite a bit more straight forward to deal with these rules
than it is to use a `genrule` all over the place.