Add `purescript_test` macro
We want to allow testing PureScript code from `bazel`.

We're taking the easy way out and leaning on `rules_nodejs` again. We
create a binary with `purescript_binary`, then call `nodejs_test` with
the binary we just created as the entry point. This is likely not the
most efficient thing to do, but it's not the worst thing in the world
either.

We can come back to this and implement a test rule properly once we have
a bit more experience and time with these rules.