Setup toolchains for PureScript binaries
We don't want to rely on binaries from the environment.

Up until now, we've been taking a simplification that both `purs` and
`purs-compile-module` would be available on the path somewhere. This is
an assumption we shouldn't make because it allows non-hermetic builds.
Instead, we can use `bazel`'s support for toolchains to bring in the
exact versions of these binaries we want.

We only support Linux and macOS at the moment, since there's no Windows
binaries for `purs-compile-module`. We'll have to distribute those at
some point. But, just to get rolling we should be alright with just
Linux and macOS.

We also add a couple of helpers to make using these toolchains easier.

At this point, we should be able to use this ruleset in the real world
and see where it needs improvement.