Setup a `gazelle` `Language` for PureScript
We want to make using the rules easy to do.

We're doing the simplest thing we can do with a lot of this stuff:
- Only generating `purescript_library` rules.
- Setting the visibility to public.
- Rolling our own parser.
- Questionably handling external dependencies.

Although there are quite a few simplifications, the important thing is,
it works! We can run something like `spago install`, have that grab the
dependencies into `.spago`, then rely on `gazelle` to traverse those
dependencies and create `BUILD.bazel` files. This isn't the most ideal
approach (since all of the external dependencies are now part of the
workspace), but it works as a proof of concept. We can refine it later.