Don't check the filename
The filename we're passed doesn't actually have an extension.

We're given the relative directory to the file, we're not given the file
itself. Since we don't have an extension, we'll never pass that
`isPureScriptFile` test.

We also don't need to check that in the grand scheme of things because
we gate any PureScript rules on the file having the right extension.