Move error conversion to error module
We want to have a simpler time with the errors.

We move the logic of how to convert from the result of calling
`Language.PureScript.Make.rebuildModule` to our errors into the error
module.

This isn't meant to be super general code. It's meant to solve a
specific problem. We might find a better name for what this does over
time. If we do, we can switch to using that name instead.

We add a couple of tests to validate this behavior.