Generate "standard" externs for `compile_module`
We want to make normal externs files other PureScript tooling can use.

The "signature" externs are super useful for `purs-compile-module`. Part
of the reason we can get cut-off is because we're using these
"signature" externs. But, they're all but useless for any other
PureScript tooling. Since all of the source annotations are null,
everything has the wrong locations.

We make a better interop story by generating the "standard" externs
files as well as the "signature" externs files at the same time. We're
not currently using them for anything in `joneshf_rules_purescript`, but
this opens the door for us to start doing stuff with other PureScript
tooling.

Closes #5