Extract externs file behavior to a module
We pull more of the externs behavior into its own module.

We follow up from pulling the "signature" externs behavior out to a
module by pulling all of the externs behavior out to a module.

This is mostly shuffling around bits without any change. Hopefully this
organization is a bit easier to deal with.