Fix dependency example with `purescript_library`
We copy/pasted too vigorously without double checking things.

Part of the reason to use `bazel` is to have it take care of a lot of
the bookkeeping. To that end, we don't actually need to reference the
transitive dependencies. In fact, it's an anti-pattern to do so. Each
rule should only list its direct dependencies, since that more
accurately describes "what to build." It's up to the rule to figure out
whatever information it needs from that information.

In any case, we don't need to list the transitive dependencies, so we
remove the "bottom of the diamond."