Only hang onto re-exported imports in "signature"
We try out a different approach to minimizing the dependencies.

We tried removing all of the imports before. The idea being that the
compiler seems to only require a transitive dependency if it's listed as
an imported module. Turns out we couldn't just remove all of the
dependencies. Any modules that are re-exported have to stay imported.
Otherwise, the compile fails with not being able to find that
re-exported module.

It sounds pretty dubious that the compiler should care about the
re-exported module after the externs have already been generated, but
that's the way it currently works. It seems like it would make more
sense to trust the externs to be right rather than not trust anything
and try to double check things every single compilation.

In any case, we can actually work around this dubious check. If we hold
onto the imports of the re-exported modules, the compiler seems to work
out okay with that information.

This has the immediate benefit that the "signature" externs are now
smaller (since there are generally fewer imports). It also has the nice
benefit of mititgating how many transitive dependencies we need in order
to compile any single module. And since we have fewer transitive
dependencies required for compiling, this has the extra nice benefit of
allowing a build system to cut-off rebuilds for some changes!

We throw together a new example that demonstrates how we can cut off
given immaterial changes. And we update the old example since we no
longer have to worry about the "base of the diamond" dependencies. With
this, we ought to have something viable for use with `bazel`.

Re: https://github.com/joneshf/purs-tools/issues/4