Set the encoding when running `hedgehog` tests
We have to work around some encoding issue on certain environments.

There's some issue with `hedgehog` tests and the encoding of
STDOUT/STDERR:
https://github.com/hedgehogqa/haskell-hedgehog/issues/110. It's not
really clear why this is always a problem with Haskell, nor is it really
important. The only thing that matters if that we have to mess around
with encodings in order to get things to work.

The `Hedgehog.Main.defaultMain` seems like it should take care of this
stuff, but it only messes with the buffering, not the encoding. In any
case, we set the encoding so we can get these tests passing in other
environments.