Setup GitHub Actions
We want to get closer to CD than we are right now.

GitHub provides actions for free for public repositories, right now, so
we take advantage of that by building artifacts for this repo.

We're setup to build just Linux and macOS since those two platforms are
straight-forward to setup. We can also build one Windows once we figure
out some of the intricacies of `bazel` and the `purescript` package.
Suffice to say, it's not nearly as straight-forward to setup.

We don't setup any crons or anything like that, but we're paving the way
to do that sort of stuff here.

We do try to cache the build artifacts in hopes that this speeds things
up. The only issue is that our cache is huge and the limit is 5GB. We
might end up thrashing the cache often.

We also upload the built binaries to the artifact storage. They're
available for 90 days there, so it ought to be long enough to do
whatever testing we need and decide if an artifact should be made into a
release (nightly or weekly).