Tar up the binary
We don't want the plain artifact, we want a tarball.

Even though the tarball consists of a single binary, we still don't want
to deal with the plain binary. We'd prefer to deal with a tarball that
can be compressed.

If we get more binaries, we might also think about tar-ing them all up
into one tarball as well as supporting them individually. For now, we
just set up the infrastructure for it to work.