Create a binary for reading module information
We want a robust way to get information about a module.

Up until now, we've been haphazardly parsing strings line-by-line to get
information out of a module. It's suprising that it's worked as well as
it has, but it's riddled with edge cases.

In order to make things a bit more robust, we make a quick little binary
that uses the actual parser to read a module and provide information
about that module. We currently provide:
- The module name.
- Any imports.
- Any re-exports.
- Whether any instances are defined.

With this information, we ought to be able to augment the `gazelle`
`language` to produce enough information that the `purescript_library`
rule can know better which modules actually need to be depended on.

We're dumping this data out as a JSON blob, but we might think about
using protobuf instead, since the support exists in both Haskell and go.
So things would be both more efficient and more robust to chang, at
least in theory.

Once we start producing binaries, we can update the rules to start using
this instead of the string parsing we were doing.