Fix nightly releases
We have to do quite a few things different from what we expected.

First, we cannot use `actions/download-artifact`. It only works within
the same workflow:
https://github.com/actions/download-artifact/issues/3. That's a bit of a
shame, since this is a supported action from GitHub. In any case, we can
use `dawidd6/action-download-artifact` instead. It allows downloading an
artifact from any workflow (or repo for that matter).

Next, we can't really figure out how to download the artifacts in one
job and make them available in a different job. There's probably a way,
but it wasn't straight-forward and we've spent enough time on this as it
is.

Next, we have to create a tag. We setup the GitHub Actions bot as the
user, since it's the thing that's kicking off the cron job. We give the
tag a nightly name and push it up. Re:
https://github.community/t/github-actions-bot-email-address/17204/5.

Finally, we have to bump up the version of `softprops/action-gh-release`
we use. The version we were on doesn't allow setting the tag name or
stuff. We want to set these things because they're not necessarily the
same as the current commit we're on. We also mark the release as a
pre-release since we don't want people to get the wrong idea. This isn't
an official release, it's just something that's available so we always
have something available without forcing people to build software
themselves.

Altogether, we've got nightlies for real now!