Handle renamed imports
We have to make sure we map up the exports to the imports.

We might have a module that imports a module as a different name and re-
exports that different name. Worse still, we might have multiple modules
imported to the same name and that one module being exported. Since
re-exports are so complex, we have to jump through a bunch of extra
hoops to find the originally imported module names.

In any case, we should now be able to deal with re-named re-exported
modules.