Decide transitive dependencies in the action
We don't want to couple rules to implementations of the toolchain.

The purpose of toolchains is to abstract away a bit of the
implementation. If the rules have to know that the underlying tool needs
all of the transitive dependencies, we've let the implementation leak
into the rules.

For the sake of having an abstraction, we move the collection of
transitive dependencies down into the actions where it's closer to the
toolchain (and has to know what the actual requirements are).

Part-of: https://github.com/joneshf/purs-tools/issues/9