Move transitive gathering closer to bundling
We want to move the transitive logic into the actions.

Similar to what we did for compilation, we want to move the transitive
dependency handling down in to the actions.

This is a bit weird because we end up creating a bunch of symlinks. We
shouldn't really have to do that, but it's due to the way the compiled
JavaScript file has its requires generated. In other words, if we were
to use a different toolchain that knew how better to handle modules, we
wouldn't necessarily have to create a ton of symlinks.

We could be able to do that by not assuming that all of the compiled
JavaScript artifacts would be "cousins" of each other (their parent
directories share a common parent). That would probably take a lot of
work to do correctly, but it could be possible.

In any case, we'll move the actual transitive stuff down into the
actions after this.