Move transitive dependency gathering into action
We don't want the rule to be so intertwined with the toolchain.

As with the `purs_compile_module` action, we want the `purs_bundle`
action to take care of gathering up any transitive dependencies
necessary to properly implement the action.

It's a little different since we're creating a bunch of symlinks. But
that's something we could do differently if we changed how we generated
each dependency. In other words, the knowledge of how dependencies work
stays in the actions rather than being in both the rules and the
actions.

We end up changing some of the tests a bit to continue to work. Instead
of giving any dependencies in the action tests, we focus on the case of
dependency-less modules. There are still rule tests that validate the
case of modules with dependencies, so we're not completely out in the
water. But we might consider testing that case explicitly in the action
tests.

Closes: https://github.com/joneshf/purs-tools/issues/9