Use the absolute path for files
We want to be able to locate the file no matter where we're run from.

This is largely necessitated by wanting to update the version of
`gazelle` we use. In the version we use, `gazelle` changes the directory
to where it's called and run from there. In newer versions, no directory
change takes place. So `gazelle` ends up being called from some bazel
directory.

Instead of assuming we're going to always be called from some workspace-
relative directory (which we can then use the relative location for), we
switch to using the absolute path and going from there.