Bump `gazelle` to a version that supports runfiles
We want to be able to add a runfile to our `gazelle` `Language`.

The previous version of `gazelle` we were on didn't support runfiles at
all. The latest commit of `gazelle` (at time of this commit) supports
runfiles for the `gazelle` binary, but not for any `Language`s. The
patch we apply adds supports for runfiles for `Language`s. This is
largely just the diff of
https://github.com/bazelbuild/bazel-gazelle/pull/828 applied to the
latest commit.

In any case, we ought to be able to now use runfiles in our `Language`.