Add `purs-module-information` to the toolchain
We want to be able to use this binary in our `gazelle` `Language`.

We need a more robust way to get information about a PureScript module
than mucking around with strings on a line-by-line basis. We made a
binary that can parse a PureScript module and provide entirely static
information about that module (i.e. no compilation required).

We add this binary to the toolchain so it's available to us (in a sort
of round-about way). We ought to be able to make use of this by forking
out to the binary to do the work and parsing its results.