Depend on `purs-module-information` in `Language`
We hook up everything and make sure we fail early if it doesn't exist.

We want to use this binary from the `Language` we're implementing, so we
make it a runfile that we need, find it in the runfiles, and hold onto
it for later. We should be able to execute the binary and get back the
module information.