Use `purs-module-information` to parse modules
This should allow us to parse any valid PureScript module.

Since we're using the underlying parser of the official compiler, we
should have enough confidence that we're going to be able to parse any
valid PureScript module. We can now drop our makeshift parser that we
wrote.

We change some fields around so we can decode the JSON output. Other
than that, we're just calling out to the binary. We'll expand our model
more to incorporate more information we get back.