Allow setting the re-exports
We want to keep track of which modules are re-exported.

Since re-exports behave differently than other imported modules, we want
to know which ones we have to do something extra for. The idea with
allowing each target to specify its re-exports is that we need this
information at analysis time. We cannot parse the file to figure it out
during analysis time. So we've gotta require it up front.

We don't use the information at the moment, but we do allow it to be
set. In that sense, it's not really required. But it'd best if we start
computing it in the `gazelle` `Language`.