Allow setting whether a module defines instances
We want to keep track of which modules are defining instances.

Typeclass instances throw a whole new wrench into the cut-off situation.
Since instances are passed implicitly through modules, there's no way to
tell statically whether a module relies on an instance. The only way to
know for sure is to compile the module with all of its dependencies and
see if it ended up needing an instance.

When you take a step back and think about it, it's quite the wonder that
this is an acceptable language feature. It also makes it much clearer
why people have a hard time learning a language like PureScript: there's
a ton of magic in typeclasses, and relying heavily on them makes
everything more complicated.

In any case, we have to deal with them. So we need to know which modules
define instances, so we can make sure they're in scope later on.