Provide `Error` as its own `haskell_library`
We can make depending on the `Error` module a bit easier.

The nice thing about `bazel` is that we can make different targets
really easily. Much easier than we could do with `cabal` or `stack`.
It's not impossible with `cabal` or `stack`, but it'd definitely be more
work to do it.

Here we're saying that `Error` is part of its own `haskell_library`
named `error`, as well as being part of the `compile-module`
`haskell_library`. We don't have to break any consumers of
`compile-module`, we don't have to do anything funny with re-exports, we
don't have to split the `Error` module out somewhere. None of that stuff
that we all assume is a requirement because of the tools we use is
actually necessary.

This small change is where I think a build system like `bazel` really
shines. It allows re-imagining how individual pieces of code fit
together into a bigger picture in a way that doesn't require language
support. This seems like the sort of thing that falls out of expecting
tons of packages/targets to exist and seeing how that shakes out,
rather than expecting a single package/target to exist and seeing how
that shakes out.