Add a new binary for inspecting externs files
We want some help working with externs files.

Ever since moving over to a binary encoding format, working with externs
has become much harder. When we're actively using externs and trying to
understand things about what we're generating, it's a bit easier if we
can see what's inside them in a decent format.

We add a binary that opens up the externs and prints information out in
a human-readable format. We only really care about instances at the
moment, so we support printing those out fairly in-depth.

We also setup things so it builds in our CD pipeline. It would be nice
to not have to remember to do this, but this seems fine for now.