Support GHC RTS options
We provide sensible defaults of limiting to a single capability.

Part of the reason using `purs-compile-module` is slower than using
`purs compile` is that `purs-compile-module` is compiled with maximum
capabilities.

This is a sensible default for running `purs-compile-module` directly.
When something else is scheduling multiple uses of `purs-compile-module`
on the same machine, each use reaching for maximum capabilities becomes
a problem. We end up with contention between each running instance and
an overall slowness. It tends to be about 30% slower to run with maximum
capabilities.

We address this by setting a more sensible default for the rules of only
asking for a single capability. Since `bazel` will schedule each
instance of `purs-compile-module` as parallel-ly as possible, it ends up
with us getting that 30% speed-up.

We don't lose all parallelism by only asking for a single capability.
GHC's RTS can still schedule things in parallel within a single core.

Finally, since different environments have different requirements, we do
allow setting the RTS options explicitly. Hopefully, people don't
actually need to muck around with them. But it's there if the default of
a single capability isn't quite right.

We should see about allowing this setting in the `gazelle` `Language` as
compiling dependencies might need different changes as well.

Related-to: https://github.com/joneshf/purs-tools/issues/4