Normalize platform name
We want to be more consistent in how we name things.

Before, we had `linux` and `macOS` in the names. This is a bit weird
because it doesn't match either what the actual platform names are or
use a consistent casing. I.e. The platforms are `Linux` and `macOS`.

We switch to using lower case for everything, so we get `linux` and
`macos`. This should be a bit easier to deal with in the long run when
we don't have to think about the casing of platform names.