Start working on a batch compiler
We try a different approach of making `purs compile` more usable.

There's still a lot of value in using `purs-compile-module`.
Unfortunately, PureScript the language makes it really hard to do the
right thing. There's just too much implicit knowledge hidden behind the
syntax. Couple that with the fact that the ecosystem is so intertwined
with itself that there's really not much cut-off you can get if you
continue to follow the patterns of the ecosystem.

Instead of giving up, we try to work more in the batch mindset. Since we
can't really get a good amount of cut-off for individual modules, we
allow working with larger chunks. Instead of thinking about a single
module, think about multiple modules at a time. This ought to allow us
to have comparable compilation times to `purs compile`.

This should address the concern originally brought up in:
https://github.com/purescript/purescript/issues/2477. Hopefully, we
don't have to maintain this program forever and we can upstream it into
the main compiler. Since at the end of the day, the only material
difference should be that it has a single new flag.