Support pre-compiled artifacts
We want to compile a batch of PureScript files incrementally.

We make it so we can compile some group of PureScript files, then take
those pre-compiled artifacts and compile more PureScript files with
those pre-compiled artifacts.

`purs-compile` works similar to how `purs-compile-module` works, except
it works as a batch compiler. `purs-compile` also works similar to how
`purs compile` works, except it works against pre-compiled artifacts.

Hopefully, this binary works a little better with `bazel` than
`purs-compile-module`. At the very least, this provides a human-usable
solution to the long-standing issue in the upstream repo about building
against pre-compiled dependencies:
https://github.com/purescript/purescript/issues/2477.

There are some things that are unideal about the way this is
implemented. The most important one is that we cannot use
`Language.PureScript.Make.make` since it has the requirement that it
needs every single transitive source file. The reason this is an issue
is that `Language.PureScript.Make.make` is the only API (that we could
find anyway) that does a batch compilation.

What we do instead of using `Language.PureScript.Make.make` is use
`Language.PureScript.Make.rebuildModule`. This API allows passing in
externs files instead of requiring every single transitive source file.

If there were something like `Language.PureScript.Make.make` that also
accepted externs files, we could still have a batch compilation. That
problem is an optimization over what we currently have. So, we ought to
be able to transparently make things quicker.

One option to address the lack of an API we need is to re-implement
`Language.PureScript.Make.make` in our project. The only reason we don't
do that is because we want to make sure this can even be a viable
solution before getting too bogged down with optimizations.

Next step is to add some `bazel` rules so it's a bit easier to use
instead of using `genrules` all over the place.