Switch back to using the batch mode
If we try a bit harder, we can make use of the batch mode.

The core of what we're doing hasn't changed, we're still taking
pre-compiled artifacts, compiling against them, and carrying along the
pre-compile artifacts.

We have to do a bit of a dance to make sure things line up properly. We
take the externs files, convert them into very simple modules, mark them
as never needing a rebuild, pass them along, and hope for the best.

Unfortunately, the way the batch mode works means that it removes any
modules that never need a rebuild. Seems like a bug, but stranger things
have been said to be normal operation. Due to the batch mode removing
things from the cache, we have to reset the cache again once it's done.
Not the biggest issue, but it does seem like it might lead to weird
interleaving of behaviors from time to time.

In any case, we've now got performance within the wheelhouse of the
official `purs compile` command without needing to pass around all of
the transitive dependencies everywhere.