Add `purs-compile` to the toolchain
We want to be able to use `purs-compile` in rules.

We follow most of what we do for the other binaries we use in the
toolchain. Nothing is hooked up to use this, currenly. We also take a
best guess as to what the interface will actually look like. We might
have to work with it a bit more to have things line up properly. But,
this should be good enough for a start.