Flesh out `purs_compile` action
We line up (what looks like) everything for this action.

If this works out the way it says it will, this is our first rule that
doesn't need to do anything funny with the transitive dependencies! This
would be a huge win if it does work out.

For the most part, this is copying what we do for `purs_compile_module`.
The differences are in what the underlying APIs look like for the
respective binaries.