Add `purescript_package` rule
We make a rule that allows us to use `purs-compile` easily.

We define this rule for building "package"s. Since there's no actual
definition of what a "package" is, we take some liberties when what we
end up doing. Most notably, we don't restrict the files to being in some
arbitrary directory. It makes zero difference to a consumer of a
"package" where the files are located, plus it would be more work than
is reasonable to implement that restriction.

With this rule implemented, we can actually use it to show examples of
using it instead of a `genrule`; much like we did with the
`purescript_library` rule.

This is enough to make using `purs-compile` easy from `bazel`. There are
still quite a few quality-of-life improvements to make. But, we'll get
to those when we get the time.