Default to warnings failing the build
We want to provide an API closer to `purs-compile-module`.

Instead of not have any way to fail on a warning (like `purs compile`)
we follow what `purs-compile-module` does and turn warnings into errors
by default. We allow opting out of this behavior by passing a similar
`--ignore-warnings` flag.

If this binary ever gets upstreamed, we'd probably have to flip the
logic of the flag (or worse, remove it altogether). But that's an issue
for the future.