Support `ignore_warnings` in `purescript_package`
We follow suit to what `purescript_library` does.

We want to be able to tell the rule whether to treat warnings as errors
or not, so we add a `bool` attribute to encode that information. This is
largely lifted from the `purescript_library` rule (and its underlying
use of `purs_compile_module`).