Make warnings consistent with `purs compile`
We make the transition easier from `purs compile`.

If we make `purs-compile` a drop-in replacement for `purs compile`, it's
a little easier to get started with. We can just replace the ` ` with a
`-` and call it a day. Then we can incrementally use more features of
`purs-compile` when they make sense.