Bump `rules_pkg` to 0.4.0
This version has some fixes for things we've wanted for a while.

In particular, we can specify the `out` attribute correctly now. What it
was previously was a `${name}.${extension}`, which meant in practice
we'd be generating something like `purs-compile.tar.gz.tar.gz`. It seems
only by happenstance that we'd also generate a `purs-compile.tar.gz`
file we could use.

In any case, we can address these issues now.