Provide a collection of all tools
We want an easier time gathering everything together.

This is primarily to make the CD workflow a little easier. It's not
actually that useful from the outside looking in the way it's
implemented. We make a tarball that has more tarballs inside of it. What
we would want if we were distributing just that tarball would be the
underlying artifacts (not their tarballs).

In any case, this is better for us overall because it means we ought to
be able to clean up these GitHub Actions files a bit.