Use `purs-tools` in GitHub Actions
We ought to be able to use the one artifact.

We change these Actions a bit to work with the one tarball instead of a
tarball for each tool we make. This ought to make maintenance a bit
easier to work with than what we had before.

There's a good chance that the bash stuff we're doing doesn't work out
quite like we want, so this might take a little iteration.